+++ src/dimensions/lookups.py
@@ -197,7 +197,7 @@ def _df_sales_channels(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
 
     def _derive_flags(df: pd.DataFrame) -> pd.DataFrame:
         # Ensure base types
-        df["SalesChannelKey"] = df["SalesChannelKey"].astype(np.int16)
+        df["SalesChannelKey"] = df["SalesChannelKey"].astype(np.int32)
 
         # Defaults
         if "SalesChannelDescription" not in df.columns:
@@ -214,7 +214,7 @@ def _df_sales_channels(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
             df["SalesChannelCode"] = df["SalesChannel"].map(_normalize_code)
 
         if "SortOrder" not in df.columns:
-            df["SortOrder"] = df["SalesChannelKey"].astype(np.int16)
+            df["SortOrder"] = df["SalesChannelKey"].astype(np.int32)
 
         # Derive based on ChannelGroup + name heuristics
         grp = df["ChannelGroup"].astype(str).str.strip().str.lower()
@@ -255,10 +255,10 @@ def _df_sales_channels(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
             )
 
         # Hard-cast
-        df["SortOrder"] = df["SortOrder"].astype(np.int16)
+        df["SortOrder"] = df["SortOrder"].astype(np.int32)
         for c in ["IsDigital", "IsPhysical", "IsThirdParty", "IsB2B", "IsAssisted", "IsOwnedChannel", "Is24x7"]:
             df[c] = df[c].astype(bool)
-        df["TypicalFulfillmentDays"] = df["TypicalFulfillmentDays"].astype(np.int16)
+        df["TypicalFulfillmentDays"] = df["TypicalFulfillmentDays"].astype(np.int32)
 
         return df
 
@@ -379,8 +379,8 @@ def _df_loyalty_tiers(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
     required = ["LoyaltyTierKey", "LoyaltyTier", "TierRank", "PointsMultiplier"]
     override = _maybe_override_rows(dim_cfg, required_cols=required)
     if override is not None:
-        override["LoyaltyTierKey"] = override["LoyaltyTierKey"].astype(np.int16)
-        override["TierRank"] = override["TierRank"].astype(np.int16)
+        override["LoyaltyTierKey"] = override["LoyaltyTierKey"].astype(np.int32)
+        override["TierRank"] = override["TierRank"].astype(np.int32)
         override["PointsMultiplier"] = override["PointsMultiplier"].astype(float)
         return override
 
@@ -391,8 +391,8 @@ def _df_loyalty_tiers(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
         (3, "Platinum", 3, 1.20),
     ]
     df = pd.DataFrame(rows, columns=required)
-    df["LoyaltyTierKey"] = df["LoyaltyTierKey"].astype(np.int16)
-    df["TierRank"] = df["TierRank"].astype(np.int16)
+    df["LoyaltyTierKey"] = df["LoyaltyTierKey"].astype(np.int32)
+    df["TierRank"] = df["TierRank"].astype(np.int32)
     df["PointsMultiplier"] = df["PointsMultiplier"].astype(float)
     return df
 
@@ -401,7 +401,7 @@ def _df_customer_acquisition_channels(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
     required = ["CustomerAcquisitionChannelKey", "AcquisitionChannel", "ChannelGroup"]
     override = _maybe_override_rows(dim_cfg, required_cols=required)
     if override is not None:
-        override["CustomerAcquisitionChannelKey"] = override["CustomerAcquisitionChannelKey"].astype(np.int16)
+        override["CustomerAcquisitionChannelKey"] = override["CustomerAcquisitionChannelKey"].astype(np.int32)
         return override
 
     rows = [
@@ -414,7 +414,7 @@ def _df_customer_acquisition_channels(dim_cfg: Dict[str, Any]) -> pd.DataFrame:
         (7, "Affiliates", "Partner"),
     ]
     df = pd.DataFrame(rows, columns=required)
-    df["CustomerAcquisitionChannelKey"] = df["CustomerAcquisitionChannelKey"].astype(np.int16)
+    df["CustomerAcquisitionChannelKey"] = df["CustomerAcquisitionChannelKey"].astype(np.int32)
     return df
